Refactoring option normalization
