+++ cli/run/index.ts
@@ -1,7 +1,7 @@
 import { realpathSync } from 'fs';
 import relative from 'require-relative';
-import { WarningHandler } from '../../src/rollup/types';
-import mergeOptions, { GenericConfigObject } from '../../src/utils/mergeOptions';
+import { mergeOptions } from '../../src/utils/mergeOptions';
+import { GenericConfigObject } from '../../src/utils/parseOptions';
 import { getAliasName } from '../../src/utils/relativeId';
 import { handleError } from '../logging';
 import batchWarnings from './batchWarnings';
@@ -107,14 +107,8 @@ async function execute(
 	} else {
 		for (const config of configs) {
 			const warnings = batchWarnings();
-			const { inputOptions, outputOptions, optionError } = mergeOptions({
-				command,
-				config,
-				defaultOnWarnHandler: warnings.add
-			});
-			if (optionError) {
-				(inputOptions.onwarn as WarningHandler)({ code: 'UNKNOWN_OPTION', message: optionError });
-			}
+			const { inputOptions, outputOptions } = mergeOptions(config, command,
+				warnings.add);
 			if (command.stdin !== false) {
 				inputOptions.plugins!.push(stdinPlugin());
 			}
