+++ cli/run/loadConfigFile.ts
@@ -2,7 +2,7 @@ import path from 'path';
 import tc from 'turbocolor';
 import * as rollup from '../../src/node-entry';
 import { RollupBuild, RollupOutput } from '../../src/rollup/types';
-import { GenericConfigObject } from '../../src/utils/mergeOptions';
+import { GenericConfigObject } from '../../src/utils/parseOptions';
 import relativeId from '../../src/utils/relativeId';
 import { handleError, stderr } from '../logging';
 import batchWarnings from './batchWarnings';
