+++ test/function/samples/warns-for-invalid-options/main.js
@@ -1,0 +1,1 @@
+export default 42;
